Update class constructor (#13)
The device id can now also be passed as a resource in the constructor. It is automatically recognized whether the device id or the entire resource is passed.
The variable `self.instrument` is now private.

Add the functions `get_device_id` and `list_connected_tektronix_generators` to find all connected tektronix generators.

Co-authored-by: bmecke <bmecke@mail.upb.de>
Reviewed-on: https://atuin.emt.uni-paderborn.de/git/emt/tektronixsg/pulls/13
Co-authored-by: Bruno Mecke <bmecke@mail.uni-paderborn.de>
Co-committed-by: Bruno Mecke <bmecke@mail.uni-paderborn.de>